+++ crates/parser/tests/parsing_failures.rs
@@ -194,17 +194,6 @@ x =
             }
 
             #[test]
-            fn block_value_with_multiple_expressions() {
-                let source = "
-x =
-  foo: 42
-  bar: 1, 2, 3
-  baz: 99
-";
-                check_parsing_fails(source);
-            }
-
-            #[test]
             fn inline_map_without_braces() {
                 let source = "
 x = foo: 42, bar: 99,